requirements.txt
Flask==0.10.1
pysaml2==2.2.0
boto==2.35.2
